Move src/MaxText/sft to src/maxtext/trainers/post_train/sft
