+++ codecov.yml
@@ -39,7 +39,8 @@ ignore:
   - "src/MaxText/inference"
   - "src/MaxText/inference_mlperf"
   - "src/MaxText/scratch_code"
-  - "src/MaxText/distillation" # code moved to src/MaxText/trainers/post_train/distillation
+  - "src/MaxText/distillation" # code moved to src/maxtext/trainers/post_train/distillation
+  - "src/MaxText/sft" # code moved to src/maxtext/trainers/post_train/sft
 
 
 flags:
